refactor: Split up builtins module into multiple files
