+++ guppylang/std/collections/__init__.py
@@ -1,0 +1,3 @@
+from guppylang.std.collections.stack import Stack, empty_stack
+
+__all__ = ["empty_stack", "Stack"]
