+++ guppylang/std/unsupported.py
@@ -86,7 +86,7 @@ def format(x): ...
 
 
 @guppy.custom(checker=UnsupportedChecker(), higher_order_value=False)
-def forozenset(x): ...
+def frozenset(x): ...
 
 
 @guppy.custom(checker=UnsupportedChecker(), higher_order_value=False)
